+++ components/EventCard.tsx
@@ -3,7 +3,7 @@
 import { AddToCalendarButton } from "add-to-calendar-button-react";
 import Link from "next/link";
 import { DeleteButton } from "./DeleteButton";
-import { translateToHtml, getDateInfoUTC, cn } from "@/utils/utils";
+import { translateToHtml, getDateInfoUTC, cn } from "@/lib/utils";
 import { AddToCalendarButtonProps } from "@/types";
 
 type EventCardProps = {
